requirements.txt
streamlit>=1.20
pandas
numpy
sentence-transformers
faiss-cpu
scikit-learn
transformers
torch
matplotlib
python-multipart
